requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
Django==4.2.16
django-allauth==65.1.0
django-bootstrap5==24.3
django-requests==0.0.2
django-sendgrid-v5==1.2.3
idna==3.10
paho-mqtt==2.1.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2024.2
requests==2.32.3
sendgrid==6.11.0
sqlparse==0.5.1
starkbank-ecdsa==2.2.0
typing_extensions==4.12.2
urllib3==2.2.3
